Add array size to table constructor, change most 'isize' to 'int'
- Mark locations of potentially concerning casts with 'implicit cast'
  comment
- Add 'n_array' argument to 'compiler_adjust_table'
- Expose 'table_set_(hash|array)' in header
- Add 'instr_set_B'

+++ src/debug.h
@@ -33,7 +33,7 @@ debug_print_value(const Value *value);
 void
 debug_disasssemble_chunk(const Chunk *chunk);
 
-isize
-debug_disassemble_instruction(const Chunk *chunk, isize index);
+int
+debug_disassemble_instruction(const Chunk *chunk, int index);
 
 #endif // LULU_DEBUG_H
